+++ Rebus.Tests/Transport/InMem/InMemTransportFactory.cs
@@ -1,16 +1,9 @@
-﻿using NUnit.Framework;
-using Rebus.Tests.Contracts.Transports;
+﻿using Rebus.Tests.Contracts.Transports;
 using Rebus.Transport;
 using Rebus.Transport.InMem;
 
 namespace Rebus.Tests.Transport.InMem
 {
-    [TestFixture]
-    public class InMemTransportBasicSendReceive : BasicSendReceive<InMemTransportFactory> { }
-
-    [TestFixture]
-    public class InMemTransportMessageExpiration : MessageExpiration<InMemTransportFactory> { }
-
     public class InMemTransportFactory : ITransportFactory
     {
         readonly InMemNetwork _network = new InMemNetwork();