+++ Rebus.Tests/Transport/InMem/InMemTransportMessageExpiration.cs
@@ -1,0 +1,8 @@
+﻿using NUnit.Framework;
+using Rebus.Tests.Contracts.Transports;
+
+namespace Rebus.Tests.Transport.InMem
+{
+    [TestFixture]
+    public class InMemTransportMessageExpiration : MessageExpiration<InMemTransportFactory> { }
+}