+++ CHANGELOG.md
@@ -1273,12 +1273,14 @@
 
 * Fix bug in the "enforce exclusive saga access" feature that would result in an exception, if the same lock bucket got hit more than once
 
-## 6.0.0-b07
+## 6.0.0-b08
 
 * Move data bus configuration to main configurer, so now you can go `.DataBus(d => d.StoreInFileSystem(@"\\network-path\folder"))`
 * Add automatic claim check support, so you can have big messages transferred as data bus attachments by going e.g. `d.SendBigMessagesAsAttachments(bodySizeThresholdBytes: 200*1000)`
 * Move cancellation support into the backoff strategy - thanks [nativenolde]
 * Change log messages when changing number of workers to include the bus name - thanks [nativenolde]
+* Add nifty fail fast helper on options configurer: `.Options(o => o.FailFastOn<TException>())` or `.Options(o => o.FailFastOn<TException>(when: ex => ex.Message.Contains("whatever")))`
+* Change `RebusTransactionScope` so it can be used inside Rebus handlers (in case that would be handy to anyone)
 
 ---